#6 Code review comments and a bit of gradle cleanup.
The usage of checkstyle is still in progress. Due to the way the
built-in chekstyle tasks work (after build) and the inability to use the
RedundantThrowsCheck due to the lack of ability to set the checkstyle
classpath, the usage of checkstyle will have to revert to using the ant
tasks, albeit through a gradle task.